Fixed memory leak and cert-manager extension switch
