+++ ES.Kubernetes.Reflector/ES.Kubernetes.Reflector.csproj
@@ -4,6 +4,7 @@
     <TargetFramework>netcoreapp2.2</TargetFramework>
     <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
     <RuntimeIdentifiers>linux-x64;linux-musl-x64;linux-arm</RuntimeIdentifiers>
+    <ServerGarbageCollection>false</ServerGarbageCollection>
   </PropertyGroup>
 
   <ItemGroup>
