+++ helm/reflector/templates/rbac-clusterRole.yaml
@@ -18,9 +18,7 @@ rules:
   - apiGroups: ["apiextensions.k8s.io"]
     resources: ["customresourcedefinitions"]
     verbs: ["*"]
-  {{- if .Values.extensions.certManager.enabled }}
   - apiGroups: ["certmanager.k8s.io"]
     resources: ["certificates", "certificates/finalizers"]
     verbs: ["*"]
-  {{- end }}
 {{- end }}
